requirements.txt
tensorflow
mlflow
fastapi
uvicorn
dvc
pytest
requests
Pillow
